// src/enums/formEnums.ts
/**
 * @description 表单字段类型
 */
export enum FieldTypeEnum {
  INPUT = "input",
  INPUT_NUMBER = "input_number",
  INPUT_PASSWORD = "input_password",
  INPUT_TEXTAREA = "input_textarea",
  SELECT = "select",
  SELECT_MULTIPLE = "select_multiple",
  RADIO_GROUP = "radio_group",
  CHECKBOX_GROUP = "checkbox_group",
  DATEPICKER = "datepicker",
  DATEPICKER_RANGE = "datepicker_range",
  TIMEPICKER = "timepicker",
  TIMEPICKER_RANGE = "timerpicker_range",
  SWITCH = "switch",
  CASCADER = "cascader",
  AUTOCOMPLETE = "autocomplete",
}